backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=24.2
psycopg2-binary>=2.9
redis>=5.0
celery>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
